# Experiment: Remove #[rustc_box] usage in the vec![] macro