--- a/tests/ui/consts/min_const_fn/bad_const_fn_body_ice.rs
+++ b/tests/ui/consts/min_const_fn/bad_const_fn_body_ice.rs
@@ -1,7 +1,7 @@
 const fn foo(a: i32) -> Vec<i32> {
     vec![1, 2, 3]
-    //~^ ERROR allocations are not allowed
-    //~| ERROR cannot call non-const method
+    //~^ ERROR cannot call non-const
+    //~| ERROR cannot call non-const
 }
 
 fn main() {}