--- a/tests/ui/error-codes/E0010-teach.rs
+++ b/tests/ui/error-codes/E0010-teach.rs
@@ -2,6 +2,6 @@
 
 #![allow(warnings)]
 
-const CON: Vec<i32> = vec![1, 2, 3]; //~ ERROR E0010
-//~| ERROR cannot call non-const method
+const CON: Vec<i32> = vec![1, 2, 3]; //~ ERROR cannot call non-const
+//~| ERROR cannot call non-const
 fn main() {}