--- a/tests/ui/error-codes/E0010-teach.stderr
+++ b/tests/ui/error-codes/E0010-teach.stderr
@@ -1,10 +1,10 @@
-error[E0010]: allocations are not allowed in constants
+error[E0015]: cannot call non-const associated function `Box::<[i32; 3]>::new` in constants
   --> $DIR/E0010-teach.rs:5:23
    |
 LL | const CON: Vec<i32> = vec![1, 2, 3];
-   |                       ^^^^^^^^^^^^^ allocation not allowed in constants
+   |                       ^^^^^^^^^^^^^
    |
-   = note: The runtime heap is not yet available at compile-time, so no runtime heap allocations can be created.
+   = note: calls in constants are limited to constant functions, tuple structs and tuple variants
    = note: this error originates in the macro `vec` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error[E0015]: cannot call non-const method `slice::<impl [i32]>::into_vec::<std::alloc::Global>` in constants
@@ -18,5 +18,4 @@ LL | const CON: Vec<i32> = vec![1, 2, 3];
 
 error: aborting due to 2 previous errors
 
-Some errors have detailed explanations: E0010, E0015.
-For more information about an error, try `rustc --explain E0010`.
+For more information about this error, try `rustc --explain E0015`.