--- a/tests/ui/error-codes/E0010.rs
+++ b/tests/ui/error-codes/E0010.rs
@@ -1,5 +1,5 @@
 #![allow(warnings)]
 
-const CON: Vec<i32> = vec![1, 2, 3]; //~ ERROR E0010
-//~| ERROR cannot call non-const method
+const CON: Vec<i32> = vec![1, 2, 3]; //~ ERROR cannot call non-const
+//~| ERROR cannot call non-const
 fn main() {}