--- a/tests/ui/macros/vec-macro-in-pattern.rs
+++ b/tests/ui/macros/vec-macro-in-pattern.rs
@@ -4,7 +4,7 @@
 
 fn main() {
     match Some(vec![42]) {
-        Some(vec![43]) => {} //~ ERROR expected a pattern, found a function call
+        Some(vec![43]) => {} //~ ERROR expected tuple struct or tuple variant, found associated function
         //~| ERROR found associated function
         //~| ERROR usage of qualified paths in this context is experimental
         _ => {}