--- a/tests/ui/statics/check-values-constraints.rs
+++ b/tests/ui/statics/check-values-constraints.rs
@@ -79,7 +79,7 @@ static STATIC10: UnsafeStruct = UnsafeStruct;
 struct MyOwned;
 
 static STATIC11: Vec<MyOwned> = vec![MyOwned];
-//~^ ERROR allocations are not allowed in statics
+//~^ ERROR cannot call non-const
 //~^^ ERROR cannot call non-const
 
 static mut STATIC12: UnsafeStruct = UnsafeStruct;
@@ -93,28 +93,28 @@ static mut STATIC14: SafeStruct = SafeStruct {
 };
 
 static STATIC15: &'static [Vec<MyOwned>] = &[
-    vec![MyOwned], //~ ERROR allocations are not allowed in statics
+    vec![MyOwned], //~ ERROR cannot call non-const
     //~^ ERROR cannot call non-const
-    vec![MyOwned], //~ ERROR allocations are not allowed in statics
+    vec![MyOwned], //~ ERROR cannot call non-const
                    //~^ ERROR cannot call non-const
 ];
 
 static STATIC16: (&'static Vec<MyOwned>, &'static Vec<MyOwned>) = (
-    &vec![MyOwned], //~ ERROR allocations are not allowed in statics
+    &vec![MyOwned], //~ ERROR cannot call non-const
     //~^ ERROR cannot call non-const
-    &vec![MyOwned], //~ ERROR allocations are not allowed in statics
+    &vec![MyOwned], //~ ERROR cannot call non-const
                     //~^ ERROR cannot call non-const
 );
 
 static mut STATIC17: SafeEnum = SafeEnum::Variant1;
 
 static STATIC19: Vec<isize> = vec![3];
-//~^ ERROR allocations are not allowed in statics
+//~^ ERROR cannot call non-const
 //~^^ ERROR cannot call non-const
 
 pub fn main() {
     let y = {
-        static x: Vec<isize> = vec![3]; //~ ERROR allocations are not allowed in statics
+        static x: Vec<isize> = vec![3]; //~ ERROR cannot call non-const
         //~^ ERROR cannot call non-const
         x
         //~^ ERROR cannot move out of static