… tests/ui/attributes/rustc-box.rs
… tests/ui/attributes/rustc-box.stderr
… tests/ui/statics/check-values-constraints.stderr
… tests/ui/unpretty/box.stdout