@Nadrieril 2025-01-03
`deep_vector` is [one `vec![]` invocation with several thousand elements](https://github.com/rust-lang/rustc-perf/blob/master/collector/compile-benchmarks/deep-vector/src/main.rs). I am quite ready to ignore it, given that nothing else is affected.