@jackh726 2025-01-03
(There may already be a codegen test.)

I personally am a big ambivalent on whether it makes sense to remove it though. On one hand, it *does* reduce the specialness, but it's not enough to remove the intrinsic altogether, so it seems like an obvious compile-time win to just use it.