@Nadrieril 2025-01-03
My hope was that this was indeed enough to remove the intrinsic (as well as `thir::Expr::Box` and `Rvalue::ShallowInitBox`). The only other use is `Box::new` and my understanding is that the intrinsic is only useful to avoid allocating a temporary, but in `Box::new` the value is already in a local so there's no need.