@RalfJung 2025-01-03
> My hope was that this was indeed enough to remove the intrinsic (as well as `thir::Expr::Box` and `Rvalue::ShallowInitBox`). The only other use is `Box::new` and my understanding is that the intrinsic is only useful to avoid allocating a temporary, but in `Box::new` the value is already in a local so there's no need.

If that is the goal then the PR should also remove that other use and we can see if the intrinsic can truly be avoided.

That said, the perf run clearly shows that the comment in the code is still accurate. So I'm not convinced we actually want to remove it. Do we have any idea where all that extra compilation time is spent and whether we can avoid it?