@saethlin 2025-01-04
> Probably worth a codegen test if we do want to remove this. Given that the difference for deep-vector is in the backend portion, my hope is that it optimizes to the same thing but just takes more time.

A codegen test of _what_ though? The compile time regression that's reported above is to unoptimized builds. As far as I can tell, LLVM is indeed fixing up our mess. Just slowly (as is tradition).

> Do we have any idea where all that extra compilation time is spent and whether we can avoid it?

Diffing the LLVM IR on nightly and on this PR, I see these changes to `main`, and no significant changes elsewhere:
```diff
 start:
+  %_4 = alloca [543864 x i8], align 4
   %_x = alloca [24 x i8], align 8
-; call alloc::alloc::exchange_malloc
-  %_5 = call ptr @_ZN5alloc5alloc15exchange_malloc17h373738c909339ef5E(i64 543864, i64 4)
-  %_9 = ptrtoint ptr %_5 to i64
-  %_12 = and i64 %_9, 3
-  %_13 = icmp eq i64 %_12, 0
-  br i1 %_13, label %bb4, label %panic
-
-bb4:                                              ; preds = %start
```
```diff
+; call alloc::boxed::Box<T>::new
+  %_3 = call align 4 ptr @"_ZN5alloc5boxed12Box$LT$T$GT$3new17h0eae96c7f92e5aa7E"(ptr align 4 %_4)
 ; call alloc::slice::<impl [T]>::into_vec
-  call void @"_ZN5alloc5slice29_$LT$impl$u20$$u5b$T$u5d$$GT$8into_vec17hf307cc899eb61d2eE"(ptr sret([24 x i8]) align 8 %_x, ptr
 align 4 %_5, i64 135966)
+  call void @"_ZN5alloc5slice29_$LT$impl$u20$$u5b$T$u5d$$GT$8into_vec17hac0528ae86e71898E"(ptr sret([24 x i8]) align 8 %_x, ptr
 align 4 %_3, i64 135966)
 ; call core::ptr::drop_in_place<alloc::vec::Vec<i32>>
-  call void @"_ZN4core3ptr47drop_in_place$LT$alloc..vec..Vec$LT$i32$GT$$GT$17h6ba5100bfd6008bdE"(ptr align 8 %_x)
+  call void @"_ZN4core3ptr47drop_in_place$LT$alloc..vec..Vec$LT$i32$GT$$GT$17hc0b4d783589713aaE"(ptr align 8 %_x)
   ret void
-
-panic:                                            ; preds = %start
-; call core::panicking::panic_misaligned_pointer_dereference
-  call void @_ZN4core9panicking36panic_misaligned_pointer_dereference17h956de8dcbda95434E(i64 4, i64 %_9, ptr align 8 @alloc_4e
2288cd26dda91e67d51c42946506bb) #14
-  unreachable
```

I think opt builds are unaffected partly because GVN turns this MIR:
```
        (*_8) = [const 0_i32, const 0_i32, const 0_i32, ...
```
Into this:
```
         (*_8) = [const 0_i32; 135966];
```
which we hand off to LLVM as a `memset` call instead of a hundred thousand gepi+store.

Just glancing at `perf stat`, enabling GVN with this branch brings build times from 1900 ms to 720 ms, but on nightly from 680 ms to 480 ms.

So somehow I think there's a double-whammy of both having the massive `alloca` that this PR introduces and also having the element-by-element initialization code that's causing the slowdown.

But simply having the big alloca with this PR is a fair point against this PR; as it stands this change will increase the stack usage of some programs and probably make them stack overflow in dev builds where they didn't before. Though this doesn't extend to much more reasonable code like `vec![0; 100_000]` because that calls `vec::from_elem` and never has a big local array.